fixing mask projection

+++ util/utilIO.cpp
@@ -112,8 +112,9 @@ cv::Mat ImgIO::projChngMaskTo3D(const cv::Mat &chngMask, const vcg::Shot<float> 
   std::cout<<"Rt 1:\n"<<cam1_Rt<<"\n intr 1:\n"<<cam1_intr<<std::endl;
   
   cv::perspectiveTransform(cam1_points, cam2_points, H);  
-  cv::Mat pnts3D(4, cam1_points.size(), CV_64FC4);
-  cv::Mat pnts3D_final(4, cam1_points.size(), CV_64FC4);
+  cv::Mat pnts3D(4, cam1_points.size(), CV_32FC4);
+
+  cv::Mat pnts3D_final(4, cam1_points.size(), CV_32FC4);
 
   cv::triangulatePoints(cam1_fmat, cam2_fmat, cv::Mat(cam1_points).reshape(1,2), cv::Mat(cam2_points).reshape(1,2), pnts3D);
